#300 - refactor Type IT packages
