+++ src/benchmark/java/com/commercetools/sync/benchmark/TypeSyncBenchmark.java
@@ -39,9 +39,9 @@ import static com.commercetools.sync.benchmark.BenchmarkUtils.calculateDiff;
 import static com.commercetools.sync.benchmark.BenchmarkUtils.saveNewResult;
 import static com.commercetools.sync.integration.commons.utils.ITUtils.deleteTypesFromTargetAndSource;
 import static com.commercetools.sync.integration.commons.utils.SphereClientUtils.CTP_TARGET_CLIENT;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.FIELD_DEFINITION_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.FIELD_DEFINITION_NAME_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.deleteTypes;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.FIELD_DEFINITION_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.FIELD_DEFINITION_NAME_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.deleteTypes;
 import static com.commercetools.tests.utils.CompletionStageUtil.executeBlocking;
 import static java.lang.String.format;
 import static java.util.Collections.singletonList;
