+++ src/integration-test/java/com/commercetools/sync/integration/externalsource/types/TypeSyncIT.java
@@ -1,4 +1,4 @@
-package com.commercetools.sync.integration.types;
+package com.commercetools.sync.integration.externalsource.types;
 
 
 import com.commercetools.sync.commons.asserts.statistics.AssertionsForStatistics;
@@ -44,20 +44,20 @@ import java.util.stream.IntStream;
 import static com.commercetools.sync.integration.commons.utils.ITUtils.deleteTypesFromTargetAndSource;
 import static com.commercetools.sync.integration.commons.utils.SphereClientUtils.CTP_SOURCE_CLIENT;
 import static com.commercetools.sync.integration.commons.utils.SphereClientUtils.CTP_TARGET_CLIENT;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.FIELD_DEFINITION_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.FIELD_DEFINITION_2;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.FIELD_DEFINITION_3;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.FIELD_DEFINITION_LABEL_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.FIELD_DEFINITION_NAME_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.TYPE_DESCRIPTION_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.TYPE_DESCRIPTION_2;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.TYPE_KEY_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.TYPE_KEY_2;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.TYPE_NAME_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.TYPE_NAME_2;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.getTypeByKey;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.populateSourceProject;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.populateTargetProject;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.FIELD_DEFINITION_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.FIELD_DEFINITION_2;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.FIELD_DEFINITION_3;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.FIELD_DEFINITION_LABEL_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.FIELD_DEFINITION_NAME_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.TYPE_DESCRIPTION_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.TYPE_DESCRIPTION_2;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.TYPE_KEY_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.TYPE_KEY_2;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.TYPE_NAME_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.TYPE_NAME_2;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.getTypeByKey;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.populateSourceProject;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.populateTargetProject;
 import static java.lang.String.format;
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
