+++ src/integration-test/java/com/commercetools/sync/integration/services/impl/TypeServiceImplIT.java
@@ -34,10 +34,10 @@ import java.util.Set;
 import static com.commercetools.sync.integration.commons.utils.CategoryITUtils.createCategoriesCustomType;
 import static com.commercetools.sync.integration.commons.utils.ITUtils.deleteTypesFromTargetAndSource;
 import static com.commercetools.sync.integration.commons.utils.SphereClientUtils.CTP_TARGET_CLIENT;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.FIELD_DEFINITION_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.TYPE_DESCRIPTION_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.TYPE_KEY_1;
-import static com.commercetools.sync.integration.types.utils.TypeITUtils.TYPE_NAME_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.FIELD_DEFINITION_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.TYPE_DESCRIPTION_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.TYPE_KEY_1;
+import static com.commercetools.sync.integration.commons.utils.TypeITUtils.TYPE_NAME_1;
 import static com.commercetools.tests.utils.CompletionStageUtil.executeBlocking;
 import static java.lang.String.format;
 import static java.util.Collections.singleton;
